[CS] decouple PropertyNaming
